Corrected missing packet handlers on client/server.
In Abhi's commit -
https://github.com/jcsnider/Orion-Game-Engine/commit/bddc2bb45691fee176d7bfeb30afe9f1b48bb5bf

He switched the packet handling to a callback system instead of the
select/switch statement. There were several packet headers that he
forgot or mismatched. This corrects most if not all of them.

This resolves issues with the item editor not saving, the resource
editor not working along with a ton of other various issues.