Orion+ version 2
- Swapped entire file handler from old VB6-compatiblity
to ArchaicSoft IO

- Renamed all client auto-copy files to capitols for cleanliness.

- removed map - preloader.
The above was buggy and cleanup is still faster than
working preloader.

- Changed some file names to logical things for the clean
look of the source.

- Changed 32bit/64bit separate handlers to AnyCpu.
(Now working universally).

- Moved Source code to a Source folder and renamed redundant Orion +
to just the object name.

- Changed build location to Build folder at .sln level for ease of
access of files.

- Changed AutoTile array in maprec to a single byte property in
tiledatarec. (Optimization)

- Optimized LOADS of data by removing complete usage of longs
(Please note to future users, Longs are massively bigger than needed
and Integers in .net are MORE than large enough for all the data
eclipse based engines EVER used!)

- Removed CheckDirectories in Client as the files are auto copied
from source and should never NOT exist unless client needs
redownloaded anyway.(It was already trying to call GUI and other
things before the CheckDirectories had a chance to verify anyway.)

- Removed useless vb6 leftover code like VKey and NLng and other
unused garbage.

- Removed Some unused code.

- Replaced blocks of Constant (0 to X) with Enums to save ram.

- Removed useless Yes No boolean constants.

- Moved Shared code to a separate Shared Library to ensure code
that has to be the same on both sides is not accidently missed.