Updated ServerLoop
Linq-ified a lot of the serverloop, it no longer loops through ALL
players for every little check, but will simply retrieve a single list
of online players at the start and use that throughout.

I've also simplified a lot of the nested if if if if nonsense into Linq
queries, hopefully this makes everything more legible.

Removed the housing timer, as it was the exact same as another and moved
it to the appropriate location.

Commented out a second occurance of Tmr25, it's nested inside another
timer and I really don't see the point of it, put it back if it ends up
being required.